refactor: move NavigationTheme and ModalTheme to theme file
Move MagicStarterNavigationTheme and MagicStarterModalTheme from
magic_starter_manager.dart to magic_starter_theme.dart to break
the circular import dependency between theme and manager files.

Both classes are pre-existing (not new to this PR), but the unified
MagicStarterTheme wrapper references them, creating theme -> manager
-> theme cycle. Now theme.dart is self-contained with no manager import.
