+++ lib/src/magic_starter_manager.dart
@@ -46,216 +46,6 @@ class MagicStarterTeamResolverConfig {
   });
 }
 
-/// Theme configuration for navigation colors and styling.
-///
-/// Allows consumer apps to override the default Wind UI `text-primary` tokens
-/// with custom colors, gradients, or light/dark mode-independent class names.
-///
-/// All fields are optional — defaults preserve the current behavior with no
-/// breaking changes.
-///
-/// ### Example
-/// ```dart
-/// MagicStarter.useNavigationTheme(
-///   MagicStarterNavigationTheme(
-///     activeItemClassName:
-///         'active:text-amber-500 active:bg-amber-500/10 dark:active:text-amber-400 dark:active:bg-amber-400/10',
-///     brandClassName:
-///         'text-lg font-bold bg-gradient-to-r from-primary-400 to-accent-500 bg-clip-text text-transparent',
-///     bottomNavActiveClassName: 'active:text-amber-500 dark:active:text-amber-400',
-///     avatarClassName: 'bg-amber-500/10 dark:bg-amber-400/10',
-///     avatarTextClassName: 'text-sm font-bold text-amber-600 dark:text-amber-400',
-///   ),
-/// );
-/// ```
-class MagicStarterNavigationTheme {
-  /// Active sidebar/drawer nav item className tokens.
-  ///
-  /// Applied to the `WDiv` that has `states: {if (isActive) 'active'}`. Each
-  /// token must include the `active:` prefix so the Wind CSS state system
-  /// activates it only when the item is selected.
-  ///
-  /// Defaults to `'active:text-primary active:bg-primary/10 dark:active:bg-primary/10'`.
-  final String activeItemClassName;
-
-  /// Hover className for sidebar/drawer nav items.
-  ///
-  /// Defaults to `'hover:bg-gray-100 dark:hover:bg-gray-800'`.
-  final String hoverItemClassName;
-
-  /// Brand/logo text className. Used when [brandBuilder] is `null`.
-  ///
-  /// Supports gradient text by combining Tailwind-like tokens, e.g.
-  /// `'text-lg font-bold bg-gradient-to-r from-primary to-accent bg-clip-text text-transparent'`.
-  ///
-  /// Defaults to `'text-lg font-bold text-primary'`.
-  final String brandClassName;
-
-  /// Custom brand/logo widget builder.
-  ///
-  /// When set, renders this widget instead of the default app name text.
-  /// Receives the current [BuildContext] and should return any widget
-  /// (image, SVG, styled text, etc.). When `null`, falls back to a [WText]
-  /// using [brandClassName].
-  ///
-  /// ```dart
-  /// brandBuilder: (context) => Image.asset('assets/logo.png', height: 28),
-  /// ```
-  final Widget Function(BuildContext context)? brandBuilder;
-
-  /// Active bottom navigation item className tokens.
-  ///
-  /// Applied to both the icon and label [WIcon]/[WText] widgets that have
-  /// `states: isActive ? {'active'} : {}`. Each token must include the
-  /// `active:` prefix.
-  ///
-  /// Defaults to `'active:text-primary'`.
-  final String bottomNavActiveClassName;
-
-  /// Avatar background className for the sidebar user menu.
-  ///
-  /// Defaults to `'bg-primary/10 dark:bg-primary/10'`.
-  final String avatarClassName;
-
-  /// Avatar text/initial color className for the sidebar user menu.
-  ///
-  /// Defaults to `'text-sm font-bold text-primary'`.
-  final String avatarTextClassName;
-
-  /// Profile dropdown trigger avatar background className.
-  ///
-  /// Used for the default circular avatar rendered in
-  /// [MagicStarterUserProfileDropdown] when no custom [triggerBuilder] is set.
-  ///
-  /// Defaults to `'bg-gradient-to-tr from-primary to-gray-200'`.
-  final String dropdownAvatarClassName;
-
-  const MagicStarterNavigationTheme({
-    this.activeItemClassName =
-        'active:text-primary active:bg-primary/10 dark:active:bg-primary/10',
-    this.hoverItemClassName = 'hover:bg-gray-100 dark:hover:bg-gray-800',
-    this.brandClassName = 'text-lg font-bold text-primary',
-    this.brandBuilder,
-    this.bottomNavActiveClassName = 'active:text-primary',
-    this.avatarClassName = 'bg-primary/10 dark:bg-primary/10',
-    this.avatarTextClassName = 'text-sm font-bold text-primary',
-    this.dropdownAvatarClassName = 'bg-gradient-to-tr from-primary to-gray-200',
-  });
-}
-
-/// Theme configuration for modal/dialog colors and styling.
-///
-/// Allows consumer apps to override the default Wind UI class names used for
-/// modal containers, headers, bodies, footers, buttons, inputs, and typography.
-///
-/// All fields are optional — defaults preserve a sensible dark-mode-aware style
-/// with no breaking changes.
-///
-/// ### Example
-/// ```dart
-/// MagicStarter.useModalTheme(
-///   MagicStarterModalTheme(
-///     containerClassName: 'bg-zinc-900 rounded-2xl border border-zinc-700',
-///     primaryButtonClassName:
-///         'px-6 py-3 rounded-xl bg-indigo-600 hover:bg-indigo-700 text-white font-semibold',
-///     maxWidth: 560.0,
-///   ),
-/// );
-/// ```
-class MagicStarterModalTheme {
-  /// Container/dialog background and border-radius className.
-  ///
-  /// Defaults to `'bg-white dark:bg-gray-800 rounded-2xl'`.
-  final String containerClassName;
-
-  /// Header section className (wraps title + description).
-  ///
-  /// Defaults to `'px-6 pt-6 pb-4'`.
-  final String headerClassName;
-
-  /// Body section className (main content area).
-  ///
-  /// Defaults to `'px-6 pb-4'`.
-  final String bodyClassName;
-
-  /// Footer section className (action buttons row).
-  ///
-  /// Defaults to `'px-6 py-4 bg-gray-50 dark:bg-gray-800/50'`.
-  final String footerClassName;
-
-  /// Title text className.
-  ///
-  /// Defaults to `'text-xl font-semibold text-gray-900 dark:text-white mb-2'`.
-  final String titleClassName;
-
-  /// Description/subtitle text className.
-  ///
-  /// Defaults to `'text-sm text-gray-600 dark:text-gray-400'`.
-  final String descriptionClassName;
-
-  /// Primary action button className.
-  ///
-  /// Defaults to
-  /// `'px-4 py-2 rounded-lg bg-primary hover:bg-primary/80 text-white text-sm font-medium'`.
-  final String primaryButtonClassName;
-
-  /// Secondary/cancel action button className.
-  ///
-  /// Defaults to
-  /// `'px-4 py-2 rounded-lg bg-white dark:bg-gray-800 border border-gray-200 dark:border-gray-700 hover:bg-gray-50 dark:hover:bg-gray-700 text-gray-700 dark:text-gray-200 text-sm font-medium'`.
-  final String secondaryButtonClassName;
-
-  /// Destructive/danger action button className.
-  ///
-  /// Defaults to
-  /// `'px-4 py-2 rounded-lg bg-red-500 hover:bg-red-600 text-white text-sm font-medium'`.
-  final String dangerButtonClassName;
-
-  /// Warning action button className.
-  ///
-  /// Defaults to
-  /// `'px-4 py-2 rounded-lg bg-amber-500 hover:bg-amber-600 text-white text-sm font-medium'`.
-  final String warningButtonClassName;
-
-  /// Inline error message text className.
-  ///
-  /// Defaults to `'text-sm text-red-600 dark:text-red-400'`.
-  final String errorClassName;
-
-  /// Text input field className.
-  ///
-  /// Defaults to
-  /// `'w-full px-3 py-3 rounded-lg bg-white dark:bg-gray-800 border border-gray-200 dark:border-gray-700 text-gray-900 dark:text-white focus:border-primary'`.
-  final String inputClassName;
-
-  /// Maximum width of the modal dialog in logical pixels.
-  ///
-  /// Defaults to `448.0`.
-  final double maxWidth;
-
-  const MagicStarterModalTheme({
-    this.containerClassName = 'bg-white dark:bg-gray-800 rounded-2xl',
-    this.headerClassName = 'px-6 pt-6 pb-4',
-    this.bodyClassName = 'px-6 pb-4',
-    this.footerClassName = 'px-6 py-4 bg-gray-50 dark:bg-gray-800/50',
-    this.titleClassName =
-        'text-xl font-semibold text-gray-900 dark:text-white mb-2',
-    this.descriptionClassName = 'text-sm text-gray-600 dark:text-gray-400',
-    this.primaryButtonClassName =
-        'px-4 py-2 rounded-lg bg-primary hover:bg-primary/80 text-white text-sm font-medium',
-    this.secondaryButtonClassName =
-        'px-4 py-2 rounded-lg bg-white dark:bg-gray-800 border border-gray-200 dark:border-gray-700 hover:bg-gray-50 dark:hover:bg-gray-700 text-gray-700 dark:text-gray-200 text-sm font-medium',
-    this.dangerButtonClassName =
-        'px-4 py-2 rounded-lg bg-red-500 hover:bg-red-600 text-white text-sm font-medium',
-    this.warningButtonClassName =
-        'px-4 py-2 rounded-lg bg-amber-500 hover:bg-amber-600 text-white text-sm font-medium',
-    this.errorClassName = 'text-sm text-red-600 dark:text-red-400',
-    this.inputClassName =
-        'w-full px-3 py-3 rounded-lg bg-white dark:bg-gray-800 border border-gray-200 dark:border-gray-700 text-gray-900 dark:text-white focus:border-primary',
-    this.maxWidth = 448.0,
-  });
-}
-
 /// Configuration for app navigation sections.
 ///
 /// Apps register these via [MagicStarter.useNavigation] to populate
